+++ library/includes/classes/class-generate-styles.php
@@ -9,7 +9,7 @@
  * @subpackage  Customize
  *
  * @since    1.8
- * @version  1.8
+ * @version  1.8.1
  *
  * Contents:
  *
@@ -961,3 +961,5 @@ final class {%= prefix_class %}_Theme_Framework_Generate_Styles {
 
 
 } // /{%= prefix_class %}_Theme_Framework_Generate_Styles
+
+add_action( 'after_setup_theme', '{%= prefix_class %}_Theme_Framework_Generate_Styles::init' );